// packages/api/src/crm/task/services/microsoftdynamicssales/types.ts
interface MicrosoftdynamicssalesTask {
    activityadditionalparams: string;
    activityid: string;
    activitytypecode: string;
    actualdurationminutes: number;
    actualend: string;
    actualstart: string;
    category: string;
    community: number;
    createdon: string;
    crmtaskassigneduniqueid: string;
    deliverylastattemptedon: string;
    deliveryprioritycode: number;
    description: string;
    descriptionblobid: string;
    descriptionblobid_name: string;
    exchangeitemid: string;
    exchangerate: number;
    exchangeweblink: string;
    formattedscheduledend: string;
    formattedscheduledstart: string;
    importsequencenumber: number;
    instancetypecode: number;
    isbilled: boolean;
    ismapiprivate: boolean;
    isregularactivity: boolean;
    isworkflowcreated: boolean;
    lastonholdtime: string;
    leftvoicemail: boolean;
    modifiedon: string;
    onholdtime: number;
    overriddencreatedon: string;
    percentcomplete: number;
    postponeactivityprocessinguntil: string;
    prioritycode: number;
    processid: string;
    scheduleddurationminutes: number;
    scheduledend: string;
    scheduledstart: string;
    senton: string;
    seriesid: string;
    sortdate: string;
    stageid: string;
    statecode: number;
    statuscode: number;
    subcategory: string;
    subject: string;
    subscriptionid: string;
    timezoneruleversionnumber: number;
    traversedpath: string;
    utcconversiontimezonecode: number;
    versionnumber: number;
    _createdby_value: string;
    _createdonbehalfby_value: string;
    _modifiedby_value: string;
    _modifiedonbehalfby_value: string;
    _ownerid_value: string;
    _owningbusinessunit_value: string;
    _owningteam_value: string;
    _owninguser_value: string;
    _regardingobjectid_value: string;
    _sendermailboxid_value: string;
    _slaid_value: string;
    _slainvokedid_value: string;
    _transactioncurrencyid_value: string;
    [key: string]: any;
}

export type MicrosoftdynamicssalesTaskInput = Partial<MicrosoftdynamicssalesTask>;
export type MicrosoftdynamicssalesTaskOutput = MicrosoftdynamicssalesTaskInput;